\name{openspending.model}
\alias{openspending.model}
\title{Openspending Model}
\usage{
  openspending.model(dataset)
}
\arguments{
  \item{dataset}{the dataset you want to see the model of}
}
\description{
  Shows you the internal model of the dataset in
  Openspending - what columns do exist and how are they
  mapped to the source file
}

