\docType{package}
\name{r-openspending-package}
\alias{openspending}
\alias{r-openspending}
\alias{r-openspending-package}
\title{A R-Package for accessing Openspending}
\description{
  Openspending is a global spending data analysis and
  visualization plattform it offers a JSON based API to
  access various information of the stored data.
  \code{r-openspending} helps accessing openspending data
  from within R
}
\author{
  Michael Bauer \email{michael.bauer@okfn.org}
}
\keyword{openspending}

