\name{openspending.as.data.frame}
\alias{openspending.as.data.frame}
\title{Openspending as.data.frame}
\usage{
  openspending.as.data.frame(data)
}
\arguments{
  \item{data}{the output from openspending.aggregate}
}
\description{
  converts the output from \code{openspending.aggregate}
  into a data.frame
}

