\name{openspending.datasets}
\alias{openspending.datasets}
\title{openspending.datasets}
\usage{
  openspending.datasets(territory = NA, language = NA)
}
\arguments{
  \item{territory}{(Optional): searches for datasets in a
  specific country (2 letter ISO codes)}

  \item{language}{(Optional): searches for datasets in a
  specifig langue (2 letter code)}
}
\description{
  Accesses all Openspending Datasets
}

