\name{openspending.distinct}
\alias{openspending.distinct}
\title{Openspending Distinct}
\usage{
  openspending.distinct(dataset, dimension)
}
\arguments{
  \item{dataset}{the dataset to work on}

  \item{dimension}{the dimension you want distinct values
  from}
}
\description{
  get the distinct (unique) values in a dimension
}

